feat: Introduce the BinaryGroupByKey/Mut types
